ovs-save: add save-flows-compact option
Currently save-flows is only to save and restore flows within the same
node. However, it's also useful to store the flows, copy them elsewhere
and perform offline flow troubleshooting.

In order to facilitate that task, add a compact version of "save-flows"
that stores the script alongside the flow dumps. The output of the
command is the directory that contains all the data, which can now be
archived and moved around more easily.

Signed-off-by: Adrian Moreno <amorenoz@redhat.com>