python/ovs-ofparse: add formatting framework
Add a generic formatting framework. The framework is composed of 3
elements: FlowStyle, FlowFormatter and FlowBuffer

A FlowStyle is a class that contains information on what style to use
for each key, value and delimiter. It supports style lookups based on
keyname, value type and supports highlighting. This all comes together
as a dictionary whose values are the styles to be used. The specific
type of the style is deliverately left undefined so any kind of object
can be stored

A FlowBuffer is an object that supports appending parts of flows

A FlowFormatter is base class for formatting. It has some common
functionality to go through the parts of the flows and format them
based on the style provided by FlowStyle. Derived classes must provide
the way to build the specific style objects they understand.

The parameters to build each style is given by the config file.

Signed-off-by: Adrian Moreno <amorenoz@redhat.com>