python/ovs-ofparse: add datapath logic processing
Create a generic Datapath Tree procesing engine.
A FlowTree arranges datapath flows based on their recirc_id.

The logic is made generic as long as the elements in the tree is derived
from TreElem

Filtering is not performed on normal processing time. Instead the
filtering is performed after the tree is built in order to include all
the subtrees that has a subflow that matches the filter

Signed-off-by: Adrian Moreno <amorenoz@redhat.com>