bond: use a full bond_member to hold bals list
The way ovs_list iterator macros work is by assuming a "struct ovs_list"
is embedded in another "struct", in this case "struct bond_member"'s
field "bals_node".

In this case, we were using a free "struct ovs_list bals". This is
not too big of a problem iff we only access the "bals_node" field of the
resulting "struct bond_member*". However, we need to access tx_bytes as
well in order to sort the bals on insertion.

The compiler would detect a potential access to an arbitrary place in
the stack and it's behavoir would be undefined.

Therefore, creating a full "struct bond_members" to hold the bals list.

Signed-off-by: Adrian Moreno <amorenoz@redhat.com>