list: remove the next variable in safe loops
Apart from being safer from the point of view of potential undefined
behavior, the multi-variable version of the safe iterators have the
benefit of not requiring the declaration of an external variable
to hold the next position.

This patch removes the extra variable from the LIST_FOR_EACH_SAFE
callers.

Signed-off-by: Adrian Moreno <amorenoz@redhat.com>