hmap: use multi-variable safe iterators
Apart from being safer from the point of view of potential undefined
behavior, the multi-variable version of the safe iterators have the
benefit of not requiring the declaration of an external variable to hold
the next position.

This patch removes the NEXT variable from HMAP_FOR_ALL_SAFE macro and
all of its callers and wrappers.

Signed-off-by: Adrian Moreno <amorenoz@redhat.com>