ovs-save: add bindir to PATH
If openvswitch is not installed in the default system's path, the
ovs-save script will fail to find the tools it requires.

Fix this by adding $bindir to the PATH in this script.
In order to do this without duplicating code, move the PATH
generation code (currently in ovs-ctl.in) to ovs-lib.

Signed-off-by: Adrian Moreno <amorenoz@redhat.com>

---
v1 -> v2: rebase