ofproto: introduce live ofproto-dpif tracing
After refactoring a and exposing some xlate-trace functions, we can
add life ofproto-dpif tracing capabilities.

Live tracing is achieved through two objects:
dpif_tracing: stores the tracing configuration and its lifetime is
managed by udpif who holds a reference to it.
dpif_tracer: is created every time an upcall is to be traced and it
emits the trace result when consumed (destroyed).

For each upcall, create a dpif_tracer if:
- Packet satisfies the filter.
- It's a recirculation of a previous packet that matched the filter.

Also, expose the capability to add and display tracing configuration
through ovs-appctl.

Signed-off-by: Adrian Moreno <amorenoz@redhat.com>