ofproto: xlate: Fix explicit drops and sampling.
The decision to add or not the explicit drop action is currently based
on whether the resulting dp action list is empty or not.

This is OK for most of the cases but if IPFIX/sFlow sampling is enabled
on the bridge, it doesn't work as expected. Sampling is a passive
observability-oriented action and should not affect drop statistics.

Fix this by taking into account the size of these sampling actions.

Fixes: a13a0209750c ("userspace: Improved packet drop statistics.")
Cc: anju.thomas@ericsson.com
Signed-off-by: Adrian Moreno <amorenoz@redhat.com>