netlink-offload: Add dummy action to store ufid.
Currently the ufid is being stored in some actions, e.g: output.

When parsing, the cookie of each action that has one is overwritten. In
addition, if many of these actions exist, we unnecessarily add the ufid
multiple times.

Organize the cookie in a cleaner way by adding an empty gact at the
begining of each action list containing the ufid as cookie.

Internally, a TC_ACT_COOKIE is added to represent this empty action.

Also, act_cookie is replaced with flow_cookie as the ufid actually
represents the flow.

Signed-off-by: Adrian Moreno <amorenoz@redhat.com>