ofproto: xlate: Make flow-sampled drops explicit.
When an action set ends in a an OFP_SAMPLE action, it means the packet
will be dropped and sampled. Make the drop explicit in this case so that
drop statistics remain accurate.

This could be done outside of the sample action, i.e: "sample(...),drop"
but datapaths optimize sample actions that are found in the last
position. So, taking into account that datapaths already report when the
last sample probability fails, it is safe to put the drop inside the
sample, i.e: "sample(...,drop)".

Signed-off-by: Adrian Moreno <amorenoz@redhat.com>