netdev-linux: Use rtnetlink to update state.
RTM_GETLINK commands can block for a considerable time if RTNL lock is
very contended, slowing down the entire main thread. The way
netdev-linux currently updates the state of each netdevs is by polling,
i.e: sending a request for each netdev on every poll loop.

A more efficient approach would be to use netlink's convenient
notification mechanism, registering callbacks to update the netdev's
internal state.

OVS already has infrastructure to do that: netlink-notifier (nln) and
rtnetlink notifiers which rely on the former and only subscribe to the
RTNLGRP_LINK mcast group.
Given that netdev-linux needs to join also inet-related mcast groups,
the obvious approach would be to use a netlink-notifier (nln), create it
in netdev-linux, and call "nln_run()" on it during "netdev_linux_run()".

However, there is a small problem with this approach.

Other subsystems already rely on rtnetlink to know when interfaces have changed
(e.g: iface-notifier, route-table). When this happens, they typically trigger
some kind of reconfiguration which is very likely to query the actual
current state of netdevs. But if netdev_linux has not yet updated its
internal state (because "netdev_linux_run()" has not run yet) and we want
netdev internal state to be cached for performance, these subsystems
will reconfigure themselves using known-to-be-dirty netdev cache.

If we want to both cache netdev internal state (e.g: flags, ether_addr) and
use a callback-based approach effectively, we need to use the same
callback chain, i.e: we need to use rtnetlink to update netdev-linux
netdevs as well when links change.

The order of callbacks cannot be guaranteed so it could still be possible to
have another subsystem be notified before netdev-linux has been given a chance
to update its state. However, subsystems would likely want to aggregate
reconfigurations and will likely have their own "_run()" function to operate.

Therefore, it could be reasonalbe to assume the other subsystems would use the
callback to set a flag that is then used to trigger reconfigurations in their
own periodic "_run()" functions (this is exactly the case of the two existing
users: iface-notifier, route-table).

This patch makes netdev-linux use rtnetlink notifier for interface updates and
independent nln notifiers for address changes.

Signed-off-by: Adrian Moreno <amorenoz@redhat.com>