netdev_linux: Refactor netdev flag update.
Currently, flags from an expected-to-be-local netdev are read
using get_flags(), which is commonly called with
&netdev->ifi_flags as argument.

Refactor it into its own function and rename "update_flags" to
"modify_flags" to increase clarity.

Besides, "netdev_linux_changed" does two things: updates the state of the
internal cache to reflect the change _and_ actually changes the flags.
This strange dual-function makes many users who only want to notify a
change call the function with "&netdev->ifi_flags" as argument. Split
this behavior in two independent functions.

As an additional benefit, this patch makes "update_flags" actually
report errors when the device is not accessible.

Signed-off-by: Adrian Moreno <amorenoz@redhat.com>